Transfer func coeffs converted to long double for qf_poly compatibility

+++ transferfuncdialog.h
@@ -26,7 +26,7 @@ private:
 public:
     explicit TransferFuncDialog(QWidget *parent = 0);
 
-    void getCoeffs(QVector<float> &a, QVector<float> &b);
+    void getCoeffs(QVector<long double> &a, QVector<long double> &b);
     
 signals:
